Added fixes to OrderAPI

+++ TheatreProject.OrderAPI/Models/DTOs/CartDetailsDto.cs
@@ -12,6 +12,6 @@ public class CartDetailsDto
     public string SeatNumbers { get; set; }
     public TicketType TicketType { get; set; }
     public int Quantity { get; set; }
-    public double PricePerTicket { get; set; }
-    public double SubTotal { get; set; }
+    public decimal PricePerTicket { get; set; }
+    public decimal SubTotal { get; set; }
 }
