+++ TheatreProject.OrderAPI/Models/DTOs/OrderDetailsDto.cs
@@ -11,6 +11,6 @@ public class OrderDetailsDto
     public string SeatNumbers { get; set; }
     public TicketType TicketType { get; set; }
     public int Quantity { get; set; }
-    public double PricePerTicket { get; set; }
-    public double SubTotal { get; set; }
+    public decimal PricePerTicket { get; set; }
+    public decimal SubTotal { get; set; }
 }
